refactor: ProductList.tsx, CartDetail.tsx 컴포넌트로 분리
