+++ .prettierrc.json
@@ -1,6 +1,6 @@
 {
   "plugins": ["prettier-plugin-tailwindcss"],
-  "printWidth": 80,
+  "printWidth": 100,
   "semi": false,
   "tabWidth": 2,
   "singleQuote": true,
