+++ src/refactoring/hooks/utils/cartUtils.ts
@@ -1,84 +1,83 @@
-import { CartItem, Coupon } from '../../../types';
+import { CartItem, Coupon, Product } from '../../../types'
 
 export const calculateItemTotal = (item: CartItem) => {
-  const { price } = item.product;
-  const { quantity } = item;
+  const { price } = item.product
+  const { quantity } = item
 
   const discount = item.product.discounts.reduce((maxDiscount, d) => {
-    return quantity >= d.quantity && d.rate > maxDiscount
-      ? d.rate
-      : maxDiscount;
-  }, 0);
+    return quantity >= d.quantity && d.rate > maxDiscount ? d.rate : maxDiscount
+  }, 0)
 
-  return price * quantity * (1 - discount);
-};
+  return price * quantity * (1 - discount)
+}
 
 export const getMaxApplicableDiscount = (item: CartItem) => {
-  const { discounts } = item.product;
-  const { quantity } = item;
-  let appliedDiscount = 0;
+  const { discounts } = item.product
+  const { quantity } = item
+  let appliedDiscount = 0
 
   for (const discount of discounts) {
     if (quantity >= discount.quantity) {
-      appliedDiscount = Math.max(appliedDiscount, discount.rate);
+      appliedDiscount = Math.max(appliedDiscount, discount.rate)
     }
   }
 
-  return appliedDiscount;
-};
+  return appliedDiscount
+}
 
-export const calculateCartTotal = (
-  cart: CartItem[],
-  selectedCoupon: Coupon | null
-) => {
-  let totalBeforeDiscount = 0;
-  let totalAfterDiscount = 0;
+export const calculateCartTotal = (cart: CartItem[], selectedCoupon: Coupon | null) => {
+  let totalBeforeDiscount = 0
+  let totalAfterDiscount = 0
 
   cart.forEach((item) => {
-    const { price } = item.product;
-    const { quantity } = item;
+    const { price } = item.product
+    const { quantity } = item
 
-    totalBeforeDiscount += price * quantity;
-    totalAfterDiscount += calculateItemTotal(item);
-  });
+    totalBeforeDiscount += price * quantity
+    totalAfterDiscount += calculateItemTotal(item)
+  })
 
-  let totalDiscount = totalBeforeDiscount - totalAfterDiscount;
+  let totalDiscount = totalBeforeDiscount - totalAfterDiscount
 
   // 쿠폰 적용
   if (selectedCoupon) {
     if (selectedCoupon.discountType === 'amount') {
-      totalAfterDiscount = Math.max(
-        0,
-        totalAfterDiscount - selectedCoupon.discountValue
-      );
+      totalAfterDiscount = Math.max(0, totalAfterDiscount - selectedCoupon.discountValue)
     } else {
-      totalAfterDiscount *= 1 - selectedCoupon.discountValue / 100;
+      totalAfterDiscount *= 1 - selectedCoupon.discountValue / 100
     }
-    totalDiscount = totalBeforeDiscount - totalAfterDiscount;
+    totalDiscount = totalBeforeDiscount - totalAfterDiscount
   }
 
   return {
     totalBeforeDiscount,
     totalAfterDiscount,
     totalDiscount,
-  };
-};
+  }
+}
 
 export const updateCartItemQuantity = (
   cart: CartItem[],
   productId: string,
-  newQuantity: number
+  newQuantity: number,
 ): CartItem[] => {
   return [...cart]
     .map((item) => {
       if (item.product.id === productId) {
-        const maxQuantity = item.product.stock;
-        const updatedQuantity = Math.max(0, Math.min(newQuantity, maxQuantity));
-        return updatedQuantity > 0
-          ? { ...item, quantity: updatedQuantity }
-          : null;
+        const maxQuantity = item.product.stock
+        const updatedQuantity = Math.max(0, Math.min(newQuantity, maxQuantity))
+        return updatedQuantity > 0 ? { ...item, quantity: updatedQuantity } : null
       }
-      return item;
+      return item
     })
-    .filter((item): item is CartItem => item !== null);
-};
+    .filter((item): item is CartItem => item !== null)
+}
+
+export const getRemainingStock = (cart: CartItem[], product: Product) => {
+  const cartItem = cart.find((item) => item.product.id === product.id)
+  return product.stock - (cartItem?.quantity || 0)
+}
+
+export const getMaxDiscount = (discounts: { quantity: number; rate: number }[]) => {
+  return discounts.reduce((max, discount) => Math.max(max, discount.rate), 0)
+}
